+++ src/test/resources/dataObjDetail_QuickActions.properties
@@ -29,5 +29,3 @@ OBJ_LINKOBJ_NO_CHOICE_OBJ = This field is required.
 
 OBJ_LINKOBJ_PLACEHOLDER_KR = Select Key result from objective 
 OBJ_LINKOBJ_NO_CHOICE_KR = This field is required.
-
-
